nutri/requirements.txt
streamlit==1.31.1
pandas==2.0.1
numpy==1.23.5
scikit-learn
scikit-learn
keras==2.12.0
emoji==2.2.0
pillow
gdown==4.7.1
